refactor(mails): split 1.7k-LOC repository into per-component files (#700)
Splits src/server/lib/postgres/repositories/mails.ts (1711 LOC) into a mails/
directory scoped by server component, per Hoie's ask:

- core.ts      per-mail CRUD (save/get/mark/delete/spam), shared across the
               SMTP-receive, send, and HTTP paths
- counters.ts  atomic UID + mod-sequence reservation (mail_uid_counters)
- http.ts      web-app read/query surface (headers, delta, search, per-account
               stats, unread notifications)
- imap.ts      IMAP protocol ops (count, range FETCH, STORE flags, SEARCH, UID
               enumeration, EXPUNGE)
- index.ts     barrel re-export so the repositories/mails import path is stable

Pure move: zero logic change. The 1571 non-import code lines are byte-identical
to the original (verified by a sorted-line diff). The only additions are the
per-file import headers and the barrel. No call site changes — every consumer
imports through the unchanged repositories barrel.

The mails.test.ts source-scan guards now read the concatenated mails/*.ts
sources (sorted for deterministic regex extraction) instead of the single file.

Closes #700.
