+++ src/server/lib/postgres/repositories/mails/counters.ts
@@ -1,0 +1,208 @@
+import { logger } from "../../../logger";
+import { pool } from "../../client";
+import { ParamValue } from "../../database";
+import {
+  USER_ID,
+  UID_DOMAIN,
+  UID_ACCOUNT,
+  MODSEQ,
+  TO_ADDRESS,
+  FROM_ADDRESS,
+  SENT,
+  MAIL_UID_COUNTERS,
+  UID_KIND,
+  UID_SCOPE,
+  LAST_UID,
+} from "../../models";
+
+/**
+ * Build the atomic UID-reservation upsert.
+ *
+ * The counter row in `mail_uid_counters` holds the most recently assigned UID.
+ * The INSERT seeds it once from the current `MAX(uid)` in `mails` (via `seedSql`,
+ * which reuses $1=user_id and $4=sent), so a deployment with existing mail keeps
+ * its sequence continuous — no UID renumbering, no UIDVALIDITY churn. Every call
+ * after the first conflicts on the composite key and takes the row lock through
+ * `DO UPDATE`, returning a strictly larger value.
+ *
+ * This is what closes the receive-path race (#617): a bare `MAX(uid)+1` read in
+ * `convertMail` followed by a later INSERT is a TOCTOU — two concurrent receipts
+ * read the same max and write the same UID. Funneling assignment through this
+ * single atomic statement removes the window for every write path (receive, send,
+ * IMAP APPEND), since they all assign through the two functions below.
+ *
+ * Pure (no DB) so the SQL shape is unit-testable without intercepting the pool.
+ * `seedParams` supply any extra placeholders `seedSql` references ($5…).
+ */
+const buildReserveUidQuery = (
+  user_id: string,
+  kind: string,
+  scope: string,
+  sent: boolean,
+  seedSql: string,
+  seedParams: ParamValue[]
+): { sql: string; values: ParamValue[] } => {
+  const sql = `
+    INSERT INTO ${MAIL_UID_COUNTERS} (${USER_ID}, ${UID_KIND}, ${UID_SCOPE}, ${SENT}, ${LAST_UID})
+    VALUES ($1, $2, $3, $4, (${seedSql}))
+    ON CONFLICT (${USER_ID}, ${UID_KIND}, ${UID_SCOPE}, ${SENT})
+    DO UPDATE SET ${LAST_UID} = ${MAIL_UID_COUNTERS}.${LAST_UID} + 1
+    RETURNING ${LAST_UID} AS next_uid
+  `;
+  return { sql, values: [user_id, kind, scope, sent, ...seedParams] };
+};
+
+/** Domain-wide UID-reservation query (kind="domain", no scope). */
+export const buildDomainUidQuery = (
+  user_id: string,
+  sent: boolean
+): { sql: string; values: ParamValue[] } => {
+  const seedSql = `
+      SELECT COALESCE(MAX(${UID_DOMAIN}), 0) + 1 FROM mails
+      WHERE ${USER_ID} = $1 AND ${SENT} = $4
+    `;
+  return buildReserveUidQuery(user_id, "domain", "", sent, seedSql, []);
+};
+
+/** Per-account UID-reservation query (kind="account", scope=address). */
+export const buildAccountUidQuery = (
+  user_id: string,
+  account: string,
+  sent: boolean
+): { sql: string; values: ParamValue[] } => {
+  const addressJson = JSON.stringify([{ address: account }]);
+  const addressCondition = sent
+    ? `${FROM_ADDRESS} @> $5::jsonb`
+    : `(${TO_ADDRESS} @> $5::jsonb OR cc_address @> $5::jsonb OR bcc_address @> $5::jsonb OR envelope_to @> $5::jsonb)`;
+  const seedSql = `
+      SELECT COALESCE(MAX(${UID_ACCOUNT}), 0) + 1 FROM mails
+      WHERE ${USER_ID} = $1
+        AND ${addressCondition}
+        AND ${SENT} = $4
+    `;
+  return buildReserveUidQuery(user_id, "account", account, sent, seedSql, [
+    addressJson,
+  ]);
+};
+
+const reserveNextUid = async (query: {
+  sql: string;
+  values: ParamValue[];
+}): Promise<number> => {
+  const result = await pool.query(query.sql, query.values);
+  return parseInt(result.rows[0]?.next_uid || "1", 10);
+};
+
+// The counter is now the authoritative UID source, so a reservation failure must
+// ABORT the save — never fall back to a fabricated UID. Returning 1 here (the old
+// behavior) would assign a value colliding with the mailbox's first message, the
+// exact duplicate-UID corruption this fix removes, now via the error path. Every
+// caller aborts cleanly on a throw: SMTP receive NACKs (sender retries), the send
+// route surfaces the error, and IMAP APPEND replies `NO` (client retries).
+export const getDomainUidNext = async (
+  user_id: string,
+  sent: boolean = false
+): Promise<number> => {
+  try {
+    return await reserveNextUid(buildDomainUidQuery(user_id, sent));
+  } catch (error) {
+    logger.error("Error getting next UID", {}, error);
+    throw error;
+  }
+};
+
+export const getAccountUidNext = async (
+  user_id: string,
+  account: string,
+  sent: boolean = false
+): Promise<number> => {
+  try {
+    return await reserveNextUid(buildAccountUidQuery(user_id, account, sent));
+  } catch (error) {
+    logger.error("Error getting account UID next", {}, error);
+    throw error;
+  }
+};
+
+/**
+ * Per-user mod-sequence reservation query (CONDSTORE, RFC 7162 §3.1).
+ *
+ * Reuses the same atomic `mail_uid_counters` upsert as UID assignment — a single
+ * counter row keyed by kind="modseq" (scope="", sent=false, both unused for this
+ * kind). RFC 7162 permits one mod-sequence namespace shared across a user's
+ * mailboxes: any single mailbox still sees a strictly-increasing subsequence,
+ * which is all the RFC requires. Using the atomic INSERT … ON CONFLICT … DO
+ * UPDATE (rather than a bare `MAX(modseq)+1` read) makes concurrent flag/receipt
+ * mutations race-free, exactly as it does for UIDs (#617).
+ *
+ * The counter seeds once from the live `MAX(modseq)` across all the user's mail,
+ * so a deployment where the DEFAULT-1 backfill already set every existing row to
+ * modseq=1 gets its first reservation at 2 — strictly greater than the initial
+ * HIGHESTMODSEQ of 1. Pure (no DB) so the SQL shape is unit-testable.
+ */
+export const buildModseqQuery = (
+  user_id: string
+): { sql: string; values: ParamValue[] } => {
+  const seedSql = `
+      SELECT COALESCE(MAX(${MODSEQ}), 0) + 1 FROM mails
+      WHERE ${USER_ID} = $1
+    `;
+  return buildReserveUidQuery(user_id, "modseq", "", false, seedSql, []);
+};
+
+// Reserve the next mod-sequence for a user's next mailbox mutation. Every
+// insert/flag-change/expunge that alters a mailbox stamps the value returned
+// here so HIGHESTMODSEQ (getHighestModseq) advances monotonically.
+export const getNextModseq = async (user_id: string): Promise<number> => {
+  try {
+    return await reserveNextUid(buildModseqQuery(user_id));
+  } catch (error) {
+    logger.error("Error reserving next modseq", {}, error);
+    throw error;
+  }
+};
+
+/**
+ * HIGHESTMODSEQ for a mailbox (RFC 7162 §3.1.2.1) — the largest mod-sequence of
+ * any message routed to it. Computed on demand as `MAX(modseq)` over the same
+ * mailbox-routing predicate the UID queries use (option (a) from #607); no
+ * materialized per-mailbox counter, so nothing to keep in sync on every write.
+ *
+ * Expunged rows are INTENTIONALLY included: an EXPUNGE bumps a message's modseq
+ * before it vanishes, and HIGHESTMODSEQ must reflect that so a resyncing client
+ * (QRESYNC, later phases) detects the removal. Returns 1 for an empty mailbox
+ * (the DEFAULT-1 floor), never 0 — a 0 HIGHESTMODSEQ signals "no persistent
+ * mod-sequences", which this store does support.
+ */
+export const getHighestModseq = async (
+  user_id: string,
+  account: string | null,
+  sent: boolean
+): Promise<number> => {
+  try {
+    let sql: string;
+    let values: ParamValue[];
+    if (account === null) {
+      sql = `
+        SELECT COALESCE(MAX(${MODSEQ}), 1) AS highest FROM mails
+        WHERE ${USER_ID} = $1 AND ${SENT} = $2
+      `;
+      values = [user_id, sent];
+    } else {
+      const addressJson = JSON.stringify([{ address: account }]);
+      const addressCondition = sent
+        ? `${FROM_ADDRESS} @> $3::jsonb`
+        : `(${TO_ADDRESS} @> $3::jsonb OR cc_address @> $3::jsonb OR bcc_address @> $3::jsonb OR envelope_to @> $3::jsonb)`;
+      sql = `
+        SELECT COALESCE(MAX(${MODSEQ}), 1) AS highest FROM mails
+        WHERE ${USER_ID} = $1 AND ${SENT} = $2 AND ${addressCondition}
+      `;
+      values = [user_id, sent, addressJson];
+    }
+    const result = await pool.query(sql, values);
+    return parseInt(result.rows[0]?.highest ?? "1", 10);
+  } catch (error) {
+    logger.error("Failed to get highest modseq", {}, error);
+    return 1;
+  }
+};
